refactor(chronos_bolt): specialize family implementation
Consolidate chronos_bolt model construction and weight mapping under its owned package. Remove unused family-local duplicates while preserving the existing plugin API.
