--- python/tensorrt_model_connect/families/chronos_bolt/default_dual_profile_decoder.py
@@ -1,3 +1,0 @@
-"""Chronos-Bolt family-owned dual-profile decoder helper namespace."""
-
-__all__: list[str] = []
