--- python/tensorrt_model_connect/families/chronos_bolt/default_dual_profile_decoder_tp.py
@@ -1,3 +1,0 @@
-"""Chronos-Bolt family-owned tensor-parallel decoder helper namespace."""
-
-__all__: list[str] = []
