--- python/tensorrt_model_connect/families/chronos_bolt/graph_blocks.py
@@ -1,3 +1,0 @@
-"""Chronos-Bolt family-owned graph block helper namespace."""
-
-__all__: list[str] = []
