+++ python/tensorrt_model_connect/families/chronos_bolt/model/model.py
@@ -1,0 +1,728 @@
+"""Family-owned TensorRT model graph and utility implementation."""
+
+from __future__ import annotations
+
+
+import numpy as np
+from tensorrt_model_connect import trt_compat
+import sys
+from ..weights import (
+    _target_np_dtype,
+)
+# Graph Ops
+
+
+trt = trt_compat.get_trt()
+
+
+def _cast_back_to_trt_dtype(
+    network: trt.INetworkDefinition,
+    tensor: trt.ITensor,
+    target_dtype: trt.DataType,
+) -> trt.ITensor:
+    """Cast a tensor back to the original TRT runtime dtype after FP32 compute."""
+    if tensor.dtype == target_dtype:
+        return tensor
+    return network.add_cast(tensor, target_dtype).get_output(0)
+
+
+def add_constant(
+    network: trt.INetworkDefinition,
+    shape: tuple[int, ...],
+    values: np.ndarray,
+    dtype: np.dtype = np.float32,
+) -> trt.ITensor:
+    """Add a constant tensor in the given *dtype* (default float32)."""
+    weights = trt.Weights(np.ascontiguousarray(values, dtype=dtype))
+    layer = network.add_constant(shape, weights)
+    return layer.get_output(0)
+
+
+def add_matmul_rhs_constant(
+    network: trt.INetworkDefinition,
+    lhs: trt.ITensor,
+    lhs_width: int,
+    rhs_width: int,
+    rhs_weights: np.ndarray,
+    dtype: np.dtype = np.float32,
+) -> trt.ITensor:
+    """Matrix multiply: lhs @ rhs_constant.  rhs is [lhs_width, rhs_width]."""
+    rank = len(tuple(lhs.shape))
+    rhs_shape = (lhs_width, rhs_width) if rank <= 2 else (1,) * (rank - 2) + (lhs_width, rhs_width)
+    rhs = add_constant(
+        network,
+        rhs_shape,
+        np.asarray(rhs_weights).reshape(rhs_shape),
+        dtype=dtype,
+    )
+    rhs = _cast_back_to_trt_dtype(network, rhs, lhs.dtype)
+    mm = network.add_matrix_multiply(
+        lhs,
+        trt.MatrixOperation.NONE,
+        rhs,
+        trt.MatrixOperation.NONE,
+    )
+    return _cast_back_to_trt_dtype(network, mm.get_output(0), lhs.dtype)
+
+
+def add_bias_sum(
+    network: trt.INetworkDefinition,
+    inp: trt.ITensor,
+    width: int,
+    bias: np.ndarray,
+    dtype: np.dtype = np.float32,
+) -> trt.ITensor:
+    """Element-wise add a bias broadcast over all non-feature axes."""
+    rank = len(tuple(inp.shape))
+    bias_shape = (width,) if rank <= 1 else (1,) * (rank - 1) + (width,)
+    bias_t = add_constant(network, bias_shape, np.asarray(bias).reshape(bias_shape), dtype=dtype)
+    bias_t = _cast_back_to_trt_dtype(network, bias_t, inp.dtype)
+    s = network.add_elementwise(inp, bias_t, trt.ElementWiseOperation.SUM)
+    return _cast_back_to_trt_dtype(network, s.get_output(0), inp.dtype)
+
+
+def add_rms_norm(
+    network: trt.INetworkDefinition,
+    inp: trt.ITensor,
+    hidden_size: int,
+    gamma: np.ndarray,
+    eps_tensor: trt.ITensor,
+    dtype: np.dtype = np.float32,
+) -> trt.ITensor:
+    """RMSNorm: gamma * (x / sqrt(mean(x^2) + eps)).
+
+    FP32 precision boundary: when dtype != float32, casts to FP32 before
+    norm computation for numerical stability, then casts back.
+
+    TRT's native normalization API implements mean-centered LayerNorm, not
+    RMSNorm, so this remains a manual shared implementation.
+    """
+    need_cast = dtype != np.float32
+    output_dtype = inp.dtype
+    if need_cast:
+        inp = network.add_cast(inp, trt.float32).get_output(0)
+        eps_tensor = network.add_cast(eps_tensor, trt.float32).get_output(0)
+    sq = network.add_elementwise(inp, inp, trt.ElementWiseOperation.PROD)
+    mean = network.add_reduce(sq.get_output(0), trt.ReduceOperation.AVG, 1 << 1, keep_dims=True)
+    denom_in = network.add_elementwise(mean.get_output(0), eps_tensor, trt.ElementWiseOperation.SUM)
+    sqrt_l = network.add_unary(denom_in.get_output(0), trt.UnaryOperation.SQRT)
+    recip = network.add_unary(sqrt_l.get_output(0), trt.UnaryOperation.RECIP)
+    normalized = network.add_elementwise(inp, recip.get_output(0), trt.ElementWiseOperation.PROD)
+    gamma_t = add_constant(network, (1, hidden_size), gamma, dtype=np.float32)
+    scaled = network.add_elementwise(
+        normalized.get_output(0), gamma_t, trt.ElementWiseOperation.PROD
+    )
+    result = scaled.get_output(0)
+    if need_cast:
+        result = _cast_back_to_trt_dtype(network, result, output_dtype)
+    return result
+
+
+def add_gelu_new(
+    network: trt.INetworkDefinition,
+    inp: trt.ITensor,
+    dtype: np.dtype = np.float32,
+) -> trt.ITensor:
+    """GELU (tanh approximation): 0.5*x*(1+tanh(sqrt(2/pi)*(x+0.044715*x^3))).
+
+    Constants are cast to ``inp.dtype`` so the elementwise ops are valid in
+    a STRONGLY_TYPED network when ``inp`` is bf16 (storage np_dtype is
+    fp16, runtime trt_dtype is bfloat16) or any other non-matching combo.
+    """
+    target_dtype = inp.dtype
+    const_shape = (1,) * max(1, len(tuple(inp.shape)))
+
+    def _const(name, value):
+        c = add_constant(network, const_shape, np.array([value], dtype=np.float32), dtype=dtype)
+        return _cast_back_to_trt_dtype(network, c, target_dtype)
+
+    # x^3
+    x_sq = network.add_elementwise(inp, inp, trt.ElementWiseOperation.PROD)
+    x_cu = network.add_elementwise(x_sq.get_output(0), inp, trt.ElementWiseOperation.PROD)
+    # 0.044715 * x^3
+    coeff = _const("coeff", 0.044715)
+    scaled_cube = network.add_elementwise(x_cu.get_output(0), coeff, trt.ElementWiseOperation.PROD)
+    # x + 0.044715 * x^3
+    inner_sum = network.add_elementwise(
+        inp, scaled_cube.get_output(0), trt.ElementWiseOperation.SUM
+    )
+    # sqrt(2/pi) * (x + 0.044715 * x^3)
+    sqrt_2_over_pi = _const("sqrt_2_over_pi", np.sqrt(2.0 / np.pi))
+    tanh_arg = network.add_elementwise(
+        sqrt_2_over_pi, inner_sum.get_output(0), trt.ElementWiseOperation.PROD
+    )
+    # tanh(...)
+    tanh_l = network.add_activation(tanh_arg.get_output(0), trt.ActivationType.TANH)
+    # 1 + tanh(...)
+    one = _const("one", 1.0)
+    one_plus_tanh = network.add_elementwise(one, tanh_l.get_output(0), trt.ElementWiseOperation.SUM)
+    # 0.5 * x
+    half = _const("half", 0.5)
+    half_x = network.add_elementwise(half, inp, trt.ElementWiseOperation.PROD)
+    # 0.5 * x * (1 + tanh(...))
+    result = network.add_elementwise(
+        half_x.get_output(0), one_plus_tanh.get_output(0), trt.ElementWiseOperation.PROD
+    )
+    return result.get_output(0)
+
+
+def make_t5_relative_position_bias(
+    num_heads: int,
+    max_seq_len: int,
+    num_buckets: int = 32,
+    max_distance: int = 128,
+) -> np.ndarray:
+    """Compute T5-style relative position bias table.
+
+    Returns: [num_heads, max_seq_len, max_seq_len] float32 bias table.
+    This is baked as a constant into the TRT graph.
+    """
+
+    def _relative_position_bucket(
+        relative_position: np.ndarray,
+        bidirectional: bool = True,
+        num_bkts: int = 32,
+        max_dist: int = 128,
+    ) -> np.ndarray:
+        """Map relative position to bucket index (T5 algorithm)."""
+        ret = np.zeros_like(relative_position, dtype=np.int32)
+        n = -relative_position
+        if bidirectional:
+            num_bkts //= 2
+            ret += (n < 0).astype(np.int32) * num_bkts
+            n = np.abs(n)
+        else:
+            n = np.maximum(n, 0)
+
+        max_exact = num_bkts // 2
+        is_small = n < max_exact
+
+        # Clamp to avoid log(0)
+        n_clamped = np.maximum(n.astype(np.float32), 1)
+        val_if_large = max_exact + (
+            np.log(n_clamped / max_exact) / np.log(max_dist / max_exact) * (num_bkts - max_exact)
+        ).astype(np.int32)
+        val_if_large = np.minimum(val_if_large, num_bkts - 1)
+
+        ret += np.where(is_small, n, val_if_large)
+        return ret
+
+    # Build relative position matrix
+    context_position = np.arange(max_seq_len, dtype=np.int32)[:, None]
+    memory_position = np.arange(max_seq_len, dtype=np.int32)[None, :]
+    relative_position = memory_position - context_position
+
+    buckets = _relative_position_bucket(
+        relative_position,
+        bidirectional=True,
+        num_bkts=num_buckets,
+        max_dist=max_distance,
+    )
+
+    return buckets.astype(np.int32)
+
+
+# Alias: add_gelu_tanh is the same as add_gelu_new (tanh approximation)
+add_gelu_tanh = add_gelu_new
+
+
+def reshape_rows_to_heads_4d(
+    network: trt.INetworkDefinition,
+    x: trt.ITensor,
+    num_heads: int,
+    head_dim: int,
+    sequence_length: int | None = None,
+    tag: str | None = None,
+) -> trt.ITensor:
+    """Reshape [S, H * D] rows into [1, H, S, D].
+
+    The transpose is required for S > 1 because each input row contains all
+    heads for one token. ``sequence_length=None`` means runtime-dynamic S.
+    """
+    seq_dim = -1 if sequence_length is None else sequence_length
+    r1 = network.add_shuffle(x)
+    if tag:
+        r1.name = tag + "_s_h_d"
+    r1.reshape_dims = (seq_dim, num_heads, head_dim)
+    r1.second_transpose = trt.Permutation([1, 0, 2])
+
+    r2 = network.add_shuffle(r1.get_output(0))
+    if tag:
+        r2.name = tag + "_1_h_s_d"
+    r2.reshape_dims = (1, num_heads, seq_dim, head_dim)
+    return r2.get_output(0)
+
+
+def reshape_heads_4d_to_rows(
+    network: trt.INetworkDefinition,
+    x_4d: trt.ITensor,
+    attention_size: int,
+    sequence_length: int | None = None,
+    tag: str | None = None,
+) -> trt.ITensor:
+    """Reshape [1, H, S, D] back to [S, H * D]."""
+    seq_dim = -1 if sequence_length is None else sequence_length
+    out = network.add_shuffle(x_4d)
+    if tag:
+        out.name = tag + "_s_h_d"
+    out.first_transpose = trt.Permutation([0, 2, 1, 3])
+    out.reshape_dims = (seq_dim, attention_size)
+    return out.get_output(0)
+
+
+def add_attention_core(
+    network: trt.INetworkDefinition,
+    q_4d: trt.ITensor,
+    k_4d: trt.ITensor,
+    v_4d: trt.ITensor,
+    causal: bool = False,
+    mask: trt.ITensor | None = None,
+    scale: float | None = None,
+    fp32_accumulation: bool = False,
+) -> trt.ITensor:
+    """Scaled dot-product attention via TRT native IAttention layer.
+
+    Replaces the manual Q@K^T → scale → softmax → @V chain.  TRT 10 fuses
+    this into a single kernel when a compatible implementation is available;
+    decomposable=True ensures a correct fallback to primitives otherwise.
+
+    NOTE: TRT IAttention computes raw BMM1 = Q @ K^T without any built-in
+    1/sqrt(D) scaling.  We pre-scale Q by 1/sqrt(D) so that the fused kernel
+    computes the standard scaled dot-product attention formula.
+
+    Args:
+        q_4d:    Query  [B, H, q_seq, D].
+        k_4d:    Key    [B, H, kv_seq, D].
+        v_4d:    Value  [B, H, kv_seq, D].
+        causal:  Apply causal (autoregressive) mask.  Mutually exclusive
+                 with ``mask``.
+        mask:    Optional additive float mask [B, H, q_seq, kv_seq] added
+                 to scaled logits before softmax.  Cannot be used with
+                 causal=True.
+        scale:   Optional Q pre-scale factor.  Defaults to 1/sqrt(D).
+        fp32_accumulation:
+                 Cast Q/K/V to FP32 before IAttention, then cast the context
+                 back to the original Q dtype.  TRT may still select a
+                 Half-input fused MHA tactic after optimizing the casts, while
+                 keeping the IAttention accumulation/output boundary in FP32.
+
+    Returns:
+        Context tensor [B, H, q_seq, D].
+    """
+    output_dtype = q_4d.dtype
+    if fp32_accumulation and output_dtype != trt.float32:
+        q_4d = network.add_cast(q_4d, trt.float32).get_output(0)
+        k_4d = network.add_cast(k_4d, trt.float32).get_output(0)
+        v_4d = network.add_cast(v_4d, trt.float32).get_output(0)
+        if mask is not None and mask.dtype != trt.float32:
+            mask = network.add_cast(mask, trt.float32).get_output(0)
+
+    # Pre-scale Q: TRT IAttention does not apply score scaling itself.
+    # Match the scale constant's dtype to Q's dtype: in strongly-typed networks
+    # a FP32 constant mixed with a FP16/BF16 Q causes add_elementwise to emit
+    # a type-mismatch error and produce a tensor with corrupted dimensions,
+    # which makes add_attention return None.
+    if scale is None:
+        head_dim = q_4d.shape[-1]
+        scale = float(1.0 / np.sqrt(head_dim)) if head_dim > 0 else 1.0
+    # Use FP16 weights directly for FP16; BF16 has no numpy native type so
+    # create as FP32 and cast; FP32 falls through to the default.
+    scale_np_dtype = np.float16 if q_4d.dtype == trt.float16 else np.float32
+    scale_t = add_constant(network, (1, 1, 1, 1), np.array([[[[scale]]]]), dtype=scale_np_dtype)
+    if q_4d.dtype == trt.bfloat16:
+        scale_t = network.add_cast(scale_t, trt.bfloat16).get_output(0)
+    q_scaled = network.add_elementwise(q_4d, scale_t, trt.ElementWiseOperation.PROD)
+
+    attn = network.add_attention(
+        q_scaled.get_output(0),
+        k_4d,
+        v_4d,
+        trt.AttentionNormalizationOp.SOFTMAX,
+        causal,
+    )
+    # Allow TRT to decompose into primitive ops when no fused kernel is
+    # available (e.g. unsupported head-dim or dtype).  This guarantees
+    # correctness on any configuration at the cost of potential performance.
+    attn.decomposable = True
+    if mask is not None and not causal:
+        attn.mask = mask
+    return _cast_back_to_trt_dtype(network, attn.get_output(0), output_dtype)
+
+
+def _scalar_constant_for_trt_dtype(
+    network: trt.INetworkDefinition,
+    shape: tuple[int, ...],
+    value: float,
+    dtype: trt.DataType,
+) -> trt.ITensor:
+    np_dtype = np.float16 if dtype == trt.float16 else np.float32
+    const = add_constant(network, shape, np.full(shape, value, dtype=np_dtype), dtype=np_dtype)
+    if dtype == trt.bfloat16:
+        const = network.add_cast(const, trt.bfloat16).get_output(0)
+    return const
+
+
+def add_tanh_softcap(
+    network: trt.INetworkDefinition,
+    tensor: trt.ITensor,
+    cap: float,
+    *,
+    scalar_shape: tuple[int, ...],
+) -> trt.ITensor:
+    """Apply ``tanh(tensor / cap) * cap`` using scalar broadcasting."""
+    cap_t = _scalar_constant_for_trt_dtype(network, scalar_shape, float(cap), tensor.dtype)
+    scaled = network.add_elementwise(tensor, cap_t, trt.ElementWiseOperation.DIV).get_output(0)
+    capped = network.add_activation(scaled, trt.ActivationType.TANH).get_output(0)
+    return network.add_elementwise(capped, cap_t, trt.ElementWiseOperation.PROD).get_output(0)
+
+
+def _repeat_kv_heads_4d(
+    network: trt.INetworkDefinition,
+    x_4d: trt.ITensor,
+    *,
+    num_heads: int,
+    num_kv_heads: int,
+    head_dim: int,
+) -> trt.ITensor:
+    if num_kv_heads == num_heads:
+        return x_4d
+    if num_kv_heads <= 0 or num_heads % num_kv_heads != 0:
+        raise ValueError(f"num_heads={num_heads} must be divisible by num_kv_heads={num_kv_heads}")
+
+    repeat = num_heads // num_kv_heads
+    if num_kv_heads == 1:
+        concat = network.add_concatenation([x_4d] * repeat)
+        concat.axis = 1
+        return concat.get_output(0)
+
+    x_shape = network.add_shape(x_4d).get_output(0)
+    one = add_constant(network, (1,), np.array([1], dtype=np.int64), dtype=np.int64)
+    seq = network.add_slice(x_shape, start=(2,), shape=(1,), stride=(1,))
+    dim = add_constant(network, (1,), np.array([head_dim], dtype=np.int64), dtype=np.int64)
+    slice_shape = network.add_concatenation([one, one, seq.get_output(0), dim])
+    slice_shape.axis = 0
+
+    repeated = []
+    for head_idx in range(num_kv_heads):
+        head_slice = network.add_slice(
+            x_4d, start=(0, head_idx, 0, 0), shape=(1, 1, 1, head_dim), stride=(1, 1, 1, 1)
+        )
+        head_slice.set_input(2, slice_shape.get_output(0))
+        repeated.extend([head_slice.get_output(0)] * repeat)
+
+    concat = network.add_concatenation(repeated)
+    concat.axis = 1
+    return concat.get_output(0)
+
+
+def _add_attention_core_with_logit_softcap(
+    network: trt.INetworkDefinition,
+    q_4d: trt.ITensor,
+    k_4d: trt.ITensor,
+    v_4d: trt.ITensor,
+    *,
+    num_heads: int,
+    num_kv_heads: int,
+    head_dim: int,
+    mask: trt.ITensor | None,
+    scale: float,
+    logit_softcap: float,
+) -> trt.ITensor:
+    output_dtype = q_4d.dtype
+    k_4d = _repeat_kv_heads_4d(
+        network, k_4d, num_heads=num_heads, num_kv_heads=num_kv_heads, head_dim=head_dim
+    )
+    v_4d = _repeat_kv_heads_4d(
+        network, v_4d, num_heads=num_heads, num_kv_heads=num_kv_heads, head_dim=head_dim
+    )
+
+    score_q = q_4d
+    score_k = k_4d
+    score_mask = mask
+    if output_dtype != trt.float32:
+        score_q = network.add_cast(score_q, trt.float32).get_output(0)
+        score_k = network.add_cast(score_k, trt.float32).get_output(0)
+        if score_mask is not None and score_mask.dtype != trt.float32:
+            score_mask = network.add_cast(score_mask, trt.float32).get_output(0)
+
+    scale_t = _scalar_constant_for_trt_dtype(network, (1, 1, 1, 1), scale, score_q.dtype)
+    scores = network.add_matrix_multiply(
+        score_q, trt.MatrixOperation.NONE, score_k, trt.MatrixOperation.TRANSPOSE
+    ).get_output(0)
+    scores = network.add_elementwise(scores, scale_t, trt.ElementWiseOperation.PROD).get_output(0)
+
+    scores = add_tanh_softcap(network, scores, logit_softcap, scalar_shape=(1, 1, 1, 1))
+
+    if score_mask is not None:
+        scores = network.add_elementwise(
+            scores, score_mask, trt.ElementWiseOperation.SUM
+        ).get_output(0)
+
+    probs = network.add_softmax(scores)
+    probs.axes = 1 << 3
+    probs_t = probs.get_output(0)
+    if probs_t.dtype != output_dtype:
+        probs_t = network.add_cast(probs_t, output_dtype).get_output(0)
+
+    context = network.add_matrix_multiply(
+        probs_t, trt.MatrixOperation.NONE, v_4d, trt.MatrixOperation.NONE
+    ).get_output(0)
+    return _cast_back_to_trt_dtype(network, context, output_dtype)
+
+
+def add_attention_from_rows(
+    network: trt.INetworkDefinition,
+    q: trt.ITensor,
+    k: trt.ITensor,
+    v: trt.ITensor,
+    *,
+    num_heads: int,
+    head_dim: int,
+    num_kv_heads: int | None = None,
+    q_seq: int | None,
+    kv_seq: int | None,
+    causal: bool = False,
+    mask: trt.ITensor | None = None,
+    scale: float | None = None,
+    logit_softcap: float | None = None,
+    fp32_accumulation: bool = False,
+    tag: str | None = None,
+) -> trt.ITensor:
+    """Native IAttention for row-major [S, H * D] Q/K/V tensors.
+
+    ``num_kv_heads`` can be smaller than ``num_heads`` for GQA/MQA. TRT
+    native IAttention supports this directly, so callers should not expand K/V
+    heads unless the model semantics require per-query-head K/V values.
+    """
+    attention_size = num_heads * head_dim
+    kv_heads = num_heads if num_kv_heads is None else num_kv_heads
+    q_4d = reshape_rows_to_heads_4d(
+        network,
+        q,
+        num_heads,
+        head_dim,
+        sequence_length=q_seq,
+        tag=None if tag is None else tag + ".q",
+    )
+    k_4d = reshape_rows_to_heads_4d(
+        network,
+        k,
+        kv_heads,
+        head_dim,
+        sequence_length=kv_seq,
+        tag=None if tag is None else tag + ".k",
+    )
+    v_4d = reshape_rows_to_heads_4d(
+        network,
+        v,
+        kv_heads,
+        head_dim,
+        sequence_length=kv_seq,
+        tag=None if tag is None else tag + ".v",
+    )
+    if scale is None:
+        scale = float(1.0 / np.sqrt(head_dim)) if head_dim > 0 else 1.0
+    if logit_softcap is not None and float(logit_softcap) > 0.0:
+        if causal:
+            raise NotImplementedError("logit_softcap attention requires an explicit additive mask")
+        ctx_4d = _add_attention_core_with_logit_softcap(
+            network,
+            q_4d,
+            k_4d,
+            v_4d,
+            num_heads=num_heads,
+            num_kv_heads=kv_heads,
+            head_dim=head_dim,
+            mask=mask,
+            scale=scale,
+            logit_softcap=float(logit_softcap),
+        )
+    else:
+        ctx_4d = add_attention_core(
+            network,
+            q_4d,
+            k_4d,
+            v_4d,
+            causal=causal,
+            mask=mask,
+            scale=scale,
+            fp32_accumulation=fp32_accumulation,
+        )
+    return reshape_heads_4d_to_rows(
+        network,
+        ctx_4d,
+        attention_size,
+        sequence_length=q_seq,
+        tag=None if tag is None else tag + ".ctx",
+    )
+
+
+# Backward-compatible name used by existing tests and call sites.
+_add_attention_core = add_attention_core
+
+
+# Time Series Trt
+
+
+trt = trt_compat.get_trt()
+
+
+def maybe_return_replicated_tp_plan(weights: dict, parallel_config) -> bytes | None:
+    if parallel_config is None or not getattr(parallel_config, "enabled", False):
+        return None
+    rank = int(getattr(parallel_config, "rank", -1))
+    if rank > 0 and "_replicated_tp_engine_plan" in weights:
+        return weights["_replicated_tp_engine_plan"]
+    return None
+
+
+def cache_replicated_tp_plan(weights: dict, parallel_config, plan: bytes) -> None:
+    if parallel_config is not None and getattr(parallel_config, "enabled", False):
+        weights["_replicated_tp_engine_plan"] = plan
+
+
+def build_serialized_network(
+    builder: trt.Builder,
+    network: trt.INetworkDefinition,
+    *,
+    precision: str,
+    verbose: bool = False,
+    tag: str = "time_series",
+) -> bytes:
+    config = builder.create_builder_config()
+    config.avg_timing_iterations = 8
+    config.max_aux_streams = 0
+    config.set_flag(trt.BuilderFlag.DISABLE_TIMING_CACHE)
+    config.set_memory_pool_limit(trt.MemoryPoolType.WORKSPACE, 4 << 30)
+    if precision == "fp32" and hasattr(trt.BuilderFlag, "TF32"):
+        config.clear_flag(trt.BuilderFlag.TF32)
+    if precision in {"fp16", "bf16"}:
+        config.set_flag(trt.BuilderFlag.FP16)
+
+    if verbose:
+        print(
+            f"[trtmc build] {tag}: building native TRT network "
+            f"({network.num_layers} layers, precision={precision}) ...",
+            file=sys.stderr,
+        )
+
+    plan = builder.build_serialized_network(network, config)
+    if plan is None:
+        raise RuntimeError(f"TensorRT {tag} engine build failed")
+    return bytes(plan)
+
+
+def create_network(*, verbose: bool = False) -> tuple[trt.Builder, trt.INetworkDefinition]:
+    logger = trt.Logger(trt.Logger.VERBOSE if verbose else trt.Logger.WARNING)
+    builder = trt.Builder(logger)
+    network = builder.create_network(
+        trt_compat.network_creation_flags(explicit_batch=True, strongly_typed=True)
+    )
+    return builder, network
+
+
+def add_linear(
+    network: trt.INetworkDefinition,
+    inp: trt.ITensor,
+    weight_out_in: np.ndarray,
+    bias: np.ndarray | None,
+    *,
+    precision: str = "fp32",
+) -> trt.ITensor:
+    target_dtype = _target_np_dtype(precision)
+    w = np.ascontiguousarray(weight_out_in.T, dtype=target_dtype)
+    out_features = int(weight_out_in.shape[0])
+    out = add_matmul_rhs_constant(
+        network,
+        inp,
+        int(weight_out_in.shape[1]),
+        out_features,
+        w,
+        dtype=target_dtype,
+    )
+    if bias is not None:
+        out = add_bias_sum(
+            network,
+            out,
+            out_features,
+            np.ascontiguousarray(bias, dtype=target_dtype),
+            dtype=target_dtype,
+        )
+    return out
+
+
+def add_scalar(
+    network: trt.INetworkDefinition,
+    shape: tuple[int, ...],
+    value: float,
+    *,
+    dtype: np.dtype = np.float32,
+) -> trt.ITensor:
+    return add_constant(
+        network,
+        shape,
+        np.full(shape, value, dtype=dtype),
+        dtype=dtype,
+    )
+
+
+def add_patchify(
+    network: trt.INetworkDefinition,
+    values: trt.ITensor,
+    *,
+    context_length: int,
+    channels: int,
+    patch_length: int,
+    patch_stride: int,
+    num_patches: int,
+) -> trt.ITensor:
+    new_sequence_length = patch_length + patch_stride * (num_patches - 1)
+    sequence_start = context_length - new_sequence_length
+    if sequence_start < 0:
+        raise ValueError("Patch configuration exceeds context length")
+
+    channel_tensors: list[trt.ITensor] = []
+    for channel in range(channels):
+        patch_tensors: list[trt.ITensor] = []
+        for patch_idx in range(num_patches):
+            start = sequence_start + patch_idx * patch_stride
+            sliced = network.add_slice(
+                values,
+                start=(0, start, channel),
+                shape=(1, patch_length, 1),
+                stride=(1, 1, 1),
+            ).get_output(0)
+            shuf = network.add_shuffle(sliced)
+            shuf.first_transpose = (0, 2, 1)
+            shuf.reshape_dims = (1, 1, 1, patch_length)
+            patch_tensors.append(shuf.get_output(0))
+        cat_patches = network.add_concatenation(patch_tensors)
+        cat_patches.axis = 2
+        channel_tensors.append(cat_patches.get_output(0))
+    cat_channels = network.add_concatenation(channel_tensors)
+    cat_channels.axis = 1
+    return cat_channels.get_output(0)
+
+
+def add_named_output(network: trt.INetworkDefinition, tensor: trt.ITensor, name: str) -> None:
+    tensor.name = name
+    network.mark_output(tensor)
+
+
+def add_gelu(network: trt.INetworkDefinition, inp: trt.ITensor) -> trt.ITensor:
+    if hasattr(trt.UnaryOperation, "ERF"):
+        inv_sqrt2 = add_scalar(
+            network, (1,) * len(tuple(inp.shape)), 1.0 / np.sqrt(2.0), dtype=np.float32
+        )
+        half = add_scalar(network, (1,) * len(tuple(inp.shape)), 0.5, dtype=np.float32)
+        one = add_scalar(network, (1,) * len(tuple(inp.shape)), 1.0, dtype=np.float32)
+        scaled = network.add_elementwise(inp, inv_sqrt2, trt.ElementWiseOperation.PROD).get_output(
+            0
+        )
+        erf = network.add_unary(scaled, trt.UnaryOperation.ERF).get_output(0)
+        one_plus = network.add_elementwise(erf, one, trt.ElementWiseOperation.SUM).get_output(0)
+        half_x = network.add_elementwise(inp, half, trt.ElementWiseOperation.PROD).get_output(0)
+        return network.add_elementwise(half_x, one_plus, trt.ElementWiseOperation.PROD).get_output(
+            0
+        )
+    return add_gelu_new(network, inp, dtype=np.float32)
