--- python/tensorrt_model_connect/families/chronos_bolt/utils.py
@@ -1,3 +1,0 @@
-"""Chronos-Bolt family-owned utility helper namespace."""
-
-__all__: list[str] = []
